[#105569758] feature-homepage-listings: refactored DealListBaseView class to a separate module
